refactor: rename persistence -> platform_store + v1 cleanup (#599, PR3/3)
Final PR of issue #599. Renames the slim control-plane store to its real role,
refreshes CLAUDE.md, and removes the last v1 stragglers.

- git mv codeframe/persistence -> codeframe/platform_store (slim control-plane
  store: auth, api keys, audit logs, interactive sessions, token usage).
- Updated all 31 codeframe.persistence imports across codeframe/, tests/, and
  scripts/. Also updated active docs: CONTRIBUTING.md, PRD.md, TESTING.md, and
  the inline python check in .github/workflows/test.yml.
- CLAUDE.md: dropped the stale `codeframe/server/` references (the dir does not
  exist after the v1 cleanup): removed the "Legacy can be read" rule and the
  matching v2-architecture bullet; replaced the `server/` line in the repo
  structure tree with `platform_store/`. Refreshed the tests/agents/ note now
  that only dependency_resolver remains.
- Test stragglers (close-out cleanup):
  - Removed orphan `sample_project_config` fixture from tests/conftest.py
    (claude-bot note from PR1; zero surviving consumers).
  - Deleted the 5 remaining #597-era v1 tests that were collect-ignored and
    referenced removed code: tests/api/{test_project_creation_api,
    test_workspace_cleanup}, tests/auth/{test_api_key_endpoints,
    test_authorization_integration, test_dual_auth}.
  - Un-ignored tests/api/test_health_endpoint.py and fixed one stale v1
    assertion (`"database" in data`) — the field was removed from the v2 /health
    response; the other five assertions match the live contract and restore
    explicit /health test coverage.
  - Emptied the root tests/conftest.py collect_ignore.
  - Deleted 3 v1 ui router tests still listed in tests/ui/conftest.py
    collect_ignore (test_deployment_mode, test_project_api, test_session_router);
    left the 2 websocket entries (pre-existing, reference live kept managers).
- Deleted tracked stale coverage artifacts (codeframe/{ui/server,platform_store/
  database}.py,cover).

Historical archives (claudedocs/, legacydocs/, specs/, sprints/, testsprite_tests/)
are intentionally untouched — they are frozen point-in-time snapshots, like
docs/archive/.

Verified: pytest --collect-only clean (3460); /health 200; cf --help; ruff +
mypy clean (181 files); tests/api/test_health_endpoint.py 3/3 pass.

Closes #599.
