+++ codeframe/auth/api_key_router.py
@@ -23,7 +23,7 @@ from codeframe.auth.api_keys import (
     SCOPE_WRITE,
 )
 from codeframe.core.api_key_service import ApiKeyService
-from codeframe.persistence.database import Database
+from codeframe.platform_store.database import Database
 
 logger = logging.getLogger(__name__)
