+++ codeframe/platform_store/__init__.py
@@ -1,5 +1,5 @@
 """Persistence layer for CodeFRAME state management."""
 
-from codeframe.persistence.database import Database
+from codeframe.platform_store.database import Database
 
 __all__ = ["Database"]
