--- tests/ui/test_project_api.py
@@ -1,127 +1,0 @@
-"""Tests for project API endpoints."""
-
-import pytest
-import os
-import tempfile
-import shutil
-from pathlib import Path
-from fastapi.testclient import TestClient
-from codeframe.persistence.database import Database
-
-from conftest import create_test_jwt_token
-
-
-@pytest.fixture
-def test_client():
-    """Create test client with temporary database and authentication."""
-    temp_dir = Path(tempfile.mkdtemp())
-    db_path = temp_dir / "test.db"
-    workspace_root = temp_dir / "workspaces"
-
-    # Save original environment
-    original_db_path = os.environ.get("DATABASE_PATH")
-    original_workspace_root = os.environ.get("WORKSPACE_ROOT")
-
-    # Set environment variables
-    os.environ["DATABASE_PATH"] = str(db_path)
-    os.environ["WORKSPACE_ROOT"] = str(workspace_root)
-
-    # Reload server module to pick up new environment
-    from codeframe.ui import server
-    from importlib import reload
-
-    reload(server)
-
-    # Initialize database
-    db = Database(db_path)
-    db.initialize()
-    server.app.state.db = db
-
-    # Create test user (user_id=1)
-    db.conn.execute(
-        """
-        INSERT OR REPLACE INTO users (
-            id, email, name, hashed_password,
-            is_active, is_superuser, is_verified, email_verified
-        )
-        VALUES (1, 'test@example.com', 'Test User', '!DISABLED!', 1, 0, 1, 1)
-        """
-    )
-    db.conn.commit()
-
-    # Initialize workspace manager
-    from codeframe.workspace import WorkspaceManager
-
-    server.app.state.workspace_manager = WorkspaceManager(workspace_root)
-
-    # Create test client with authentication headers
-    auth_token = create_test_jwt_token(user_id=1)
-    client = TestClient(server.app, headers={"Authorization": f"Bearer {auth_token}"})
-
-    yield client
-
-    # Cleanup
-    db.close()
-    shutil.rmtree(temp_dir, ignore_errors=True)
-
-    # Restore original environment
-    if original_db_path is not None:
-        os.environ["DATABASE_PATH"] = original_db_path
-    else:
-        os.environ.pop("DATABASE_PATH", None)
-
-    if original_workspace_root is not None:
-        os.environ["WORKSPACE_ROOT"] = original_workspace_root
-    else:
-        os.environ.pop("WORKSPACE_ROOT", None)
-
-
-def test_create_project_minimal(test_client):
-    """Test creating project with minimal required fields."""
-    response = test_client.post(
-        "/api/projects", json={"name": "Test Project", "description": "A test project"}
-    )
-
-    assert response.status_code == 201
-    data = response.json()
-    assert data["name"] == "Test Project"
-    # Note: API response doesn't include description/source fields currently
-
-
-def test_create_project_git_remote(test_client):
-    """Test creating project from git repository (will fail on fake URL)."""
-    response = test_client.post(
-        "/api/projects",
-        json={
-            "name": "Git Project",
-            "description": "From git",
-            "source_type": "git_remote",
-            "source_location": "https://github.com/user/repo.git",
-        },
-    )
-
-    # This will fail with 500 because the git URL is fake
-    # For now, just verify it's not a validation error (422)
-    assert response.status_code in [201, 500]  # 201 if git works, 500 if repo doesn't exist
-
-
-def test_create_project_validation_error(test_client):
-    """Test validation error for missing source_location."""
-    response = test_client.post(
-        "/api/projects",
-        json={
-            "name": "Test",
-            "description": "Test",
-            "source_type": "git_remote",
-            # Missing source_location
-        },
-    )
-
-    assert response.status_code == 422
-
-
-def test_create_project_missing_description(test_client):
-    """Test validation error for missing description."""
-    response = test_client.post("/api/projects", json={"name": "Test"})
-
-    assert response.status_code == 422
